Project restructure
- Restructured project into Python package format using Poetry.
