+++ software_engineer_assignment/gui_components.py
