+++ software_engineer_assignment/serial_interface.py
